Add files via upload
Description

This data set maps the distribution and health status of White Pines in Madison Borough, New Jersey, as of 2024. Using data from iNaturalist and NJDEPBGIS, it categorizes the trees' conditions and explores their relationship to nearby open spaces. This helps in assessing how environmental factors influence the health of White Pines in the area.